Minor tweaks

+++ PKHeX.Core.AutoMod/AutoMod/Util/InternalNetUtil.cs
@@ -57,7 +57,7 @@ public static class InternalNetUtil
     /// <param name="generation">The generation for the game the Pokémon is being uploaded from.</param>
     /// <param name="Url">location to fetch from</param>
     /// <returns></returns>
-    public static async Task<HttpResponseMessage> GPSSPost(byte[] data, int generation, string Url = "flagbrew.org")
+    public static async Task<HttpResponseMessage> GPSSPost(byte[] data, byte generation, string Url = "flagbrew.org")
     {
         using var client = new HttpClient();
